test: split scanner coverage tests by domain
